// src/app/share/enumerations/education.enum.ts
export class DemographicEducation {
    scale: EducationScale;
    value: EducationAcademic | EducationGerman;
}

export enum EducationGerman {
    HAUPTSCHULABSCHLUSS = 'hauptschulabschluss',
    REALSCHULABSCHLUSS = 'realschulabschluss',
    HOCHSCHULREIFE = 'hochschulreife',
    BACHELOR = 'bachelor',
    MASTER = 'master',
    DOKTOR = 'doktor',
    AUSBILDUNG = 'ausbildung',
    MEISTER = 'meister',
    TECHNIKER = 'techniker',
}


export enum EducationAcademic {
    BACHELOR = 'bachelor',
    MASTER = 'master',
    PHD = 'phd',
}

export enum EducationScale {
    GERMAN,
    ACADEMIC,
}